+++ 2p/srclib/utils.h
@@ -102,7 +102,7 @@ int fprintf_sq_mtx(FILE *f, uint8_t **mtx, size_t sz);
  * Utilities for deterministic finite automata
  *
 *******************************************************************/
-uint8_t dfa_delta(uint8_t ***trans_tb, uint8_t nstates, uint8_t from, uint8_t sym_idx);
+int dfa_delta(uint8_t ***trans_tb, uint8_t nstates, uint8_t from, uint8_t sym_idx);
 
 /******************************************************************
  *
